refactor(harness): move Remote/load_host into executor.py (fix inverted seam dependency)
